Split CAN-related libraries, install targets, create SDO responder
