+++ libraries/CMakeLists.txt
@@ -2,6 +2,8 @@
 # Authors: Juan G. Victores & Raul de Santos Rico
 
 add_subdirectory(CanBusSharerLib)
+add_subdirectory(CanOpenNodeLib)
 add_subdirectory(DeviceMapperLib)
 add_subdirectory(DextraRawControlboardLib)
+add_subdirectory(StateObserverLib)
 add_subdirectory(YarpPlugins)
