+++ libraries/CanBusSharerLib/ICanBusSharer.hpp
@@ -5,16 +5,22 @@
 
 #include <vector>
 
+#include "CanMessageNotifier.hpp"
 #include "CanSenderDelegate.hpp"
 
 namespace roboticslab
 {
 
 /**
+ * @defgroup CanBusSharerLib
+ * @brief Common interfaces and utilities for a network of CAN nodes.
+ */
+
+/**
  * @ingroup CanBusSharerLib
  * @brief Abstract base for a CAN bus sharer.
  */
-class ICanBusSharer
+class ICanBusSharer : public CanMessageNotifier
 {
 public:
     //! Destructor.
@@ -32,9 +38,6 @@ public:
 
     //! Finalize CAN node communications.
     virtual bool finalize() = 0;
-
-    //! Parse incoming CAN message.
-    virtual bool interpretMessage(const can_message & message) = 0;
 
     //! Pass a handle to a CAN sender delegate instance.
     virtual bool registerSender(CanSenderDelegate * sender) = 0;
