+++ libraries/YarpPlugins/CanBusControlboard/CanRxTxThreads.hpp
@@ -8,9 +8,9 @@
 #include <unordered_map>
 
 #include <yarp/os/Thread.h>
+
 #include <yarp/dev/CanBusInterface.h>
 
-#include "CanSenderDelegate.hpp"
 #include "ICanBusSharer.hpp"
 
 namespace roboticslab
@@ -82,7 +82,7 @@ private:
 
 /**
  * @ingroup CanBusControlboard
- * @brief A thread that attends CAN reads.
+ * @brief A thread that deals with CAN reads.
  *
  * Messages are forwarded to each raw subdevice given the CAN node ID.
  */
