+++ libraries/YarpPlugins/CuiAbsolute/CuiAbsolute.hpp
@@ -59,7 +59,7 @@ public:
     //  --------- ICanBusSharer declarations. Implementation in ICanBusSharerImpl.cpp ---------
 
     virtual unsigned int getId() override;
-    virtual bool interpretMessage(const can_message & message) override;
+    virtual bool notifyMessage(const can_message & message) override;
     virtual bool initialize() override;
     virtual bool finalize() override;
     virtual bool registerSender(CanSenderDelegate * sender) override;
