+++ libraries/YarpPlugins/DextraCanControlboard/DextraCanControlboard.hpp
@@ -63,7 +63,7 @@ public:
     virtual unsigned int getId() override;
     virtual bool initialize() override;
     virtual bool finalize() override;
-    virtual bool interpretMessage(const can_message & message) override;
+    virtual bool notifyMessage(const can_message & message) override;
     virtual bool registerSender(CanSenderDelegate * sender) override;
     virtual bool synchronize() override;
