+++ libraries/YarpPlugins/LacqueyFetch/LacqueyFetch.hpp
@@ -47,7 +47,7 @@ public:
     //  --------- ICanBusSharer declarations. Implementation in LacqueyFetch.cpp ---------
 
     virtual unsigned int getId() override;
-    virtual bool interpretMessage(const can_message & message) override;
+    virtual bool notifyMessage(const can_message & message) override;
     virtual bool initialize() override;
     virtual bool finalize() override;
     virtual bool registerSender(CanSenderDelegate * sender) override;
