+++ tests/CMakeLists.txt
@@ -8,7 +8,13 @@ cmake_dependent_option(ENABLE_tests "Enable/disable unit tests" ON
 if(ENABLE_tests)
     enable_testing()
     add_subdirectory(${GTestSources_SOURCE_DIR} ${CMAKE_BINARY_DIR}/gtest)
-    include_directories(${GTestSources_INCLUDE_DIR})
+    include_directories(${GTestSources_INCLUDE_DIR} ${CMAKE_CURRENT_SOURCE_DIR})
+    
+    # FutureObserverLib
+    
+    file(WRITE ${CMAKE_CURRENT_BINARY_DIR}/FutureObserverLib.cpp "#include \"FutureObserverLib.hpp\"")
+    add_library(FutureObserverLib STATIC FutureObserverLib.hpp
+                                         ${CMAKE_CURRENT_BINARY_DIR}/FutureObserverLib.cpp)
 
     # testCanBusSharerLib
 
@@ -16,6 +22,22 @@ if(ENABLE_tests)
         add_executable(testCanBusSharerLib testCanBusSharerLib.cpp)
         target_link_libraries(testCanBusSharerLib CanBusSharerLib gtest_main)
         gtest_add_tests(testCanBusSharerLib "" AUTO)
+    endif()
+    
+    # testStateObserverLib
+
+    if(TARGET StateObserverLib)
+        add_executable(testStateObserverLib testStateObserverLib.cpp)
+        target_link_libraries(testStateObserverLib StateObserverLib FutureObserverLib gtest_main)
+        gtest_add_tests(testStateObserverLib "" AUTO)
+    endif()
+    
+    # testCanOpenNodeLib
+
+    if(TARGET CanOpenNodeLib)
+        add_executable(testCanOpenNodeLib testCanOpenNodeLib.cpp)
+        target_link_libraries(testCanOpenNodeLib CanOpenNodeLib FutureObserverLib gtest_main)
+        gtest_add_tests(testCanOpenNodeLib "" AUTO)
     endif()
 
     # testDeviceMapperLib
