+++ tests/testStateObserverLib.cpp
@@ -1,0 +1,114 @@
+#include "gtest/gtest.h"
+
+#include <future>
+
+#include "StateObserver.hpp"
+
+#include "FutureObserverLib.hpp"
+
+namespace roboticslab
+{
+
+namespace test
+{
+
+/**
+ * @ingroup yarp_devices_tests
+ * @defgroup testStateObserverLib
+ * @brief Unit tests related to @ref StateObserverLib.
+ */
+
+/**
+ * @ingroup testStateObserverLib
+ * @brief See @ref FutureObserver.
+ */
+class CanBusSharerTest : public testing::Test,
+                         protected FutureObserver
+{
+public:
+    virtual void SetUp()
+    {
+    }
+
+    virtual void TearDown()
+    {
+        shutdown();
+    }
+
+protected:
+    static constexpr double TIMEOUT = 0.125; // [s]
+};
+
+TEST_F(CanBusSharerTest, StateObserver)
+{
+    // test StateObserver
+
+    StateObserver emptyStateObserver(TIMEOUT);
+    f() = std::async(std::launch::async, observer_timer{MILLIS, [&]{ return emptyStateObserver.notify(); }});
+    ASSERT_TRUE(emptyStateObserver.await());
+
+    // test TypedStateObserver<dummy> with dummy = user-defined type
+
+    struct dummy
+    { int val; };
+
+    dummy d;
+    const dummy _v1{4};
+    TypedStateObserver<dummy> dummyStateObserver(TIMEOUT);
+    f() = std::async(std::launch::async, observer_timer{MILLIS, [&]{ return dummyStateObserver.notify(_v1); }});
+    ASSERT_TRUE(dummyStateObserver.await(d));
+    ASSERT_EQ(d.val, _v1.val);
+
+    // test TypedStateObserver<int>, use notify(int)
+
+    int val1;
+    const int _v2 = 4;
+    TypedStateObserver<int> intStateObserver(TIMEOUT);
+    f() = std::async(std::launch::async, observer_timer{MILLIS, [&]{ return intStateObserver.notify(_v2); }});
+    ASSERT_TRUE(intStateObserver.await(&val1));
+    ASSERT_EQ(val1, _v2);
+
+    // test TypedStateObserver<int>, use notify(const std::uint8_t *, std::size_t)
+
+    std::int8_t val2;
+    const std::uint8_t _v3[] = {4};
+    TypedStateObserver<std::int8_t> intBuffStateObserver(TIMEOUT);
+    f() = std::async(std::launch::async, observer_timer{MILLIS, [&]{ return intBuffStateObserver.notify(_v3, 1); }});
+    ASSERT_TRUE(intBuffStateObserver.await(&val2));
+    ASSERT_EQ(val2, _v3[0]);
+
+    // test TypedStateObserver<std::uint8_t[]>
+
+    std::uint8_t buff[1];
+    const std::uint8_t _v4[] = {4};
+    TypedStateObserver<std::uint8_t[]> bufferStateObserver(TIMEOUT);
+    f() = std::async(std::launch::async, observer_timer{MILLIS, [&]{ return bufferStateObserver.notify(_v4, 1); }});
+    ASSERT_TRUE(bufferStateObserver.await(buff));
+    ASSERT_EQ(buff[0], _v4[0]);
+
+    // test StateObserver, never call notify()
+
+    StateObserver orphanStateObserver1(TIMEOUT);
+    ASSERT_FALSE(orphanStateObserver1.await());
+
+    // test StateObserver, never call await()
+
+    StateObserver orphanStateObserver2(TIMEOUT);
+    ASSERT_TRUE(orphanStateObserver2.notify());
+
+    // test StateObserver on existing instance, await() but don't notify()
+
+    ASSERT_FALSE(emptyStateObserver.await());
+
+    // test StateObserver on existing instance, await() and notify()
+
+    f() = std::async(std::launch::async, observer_timer{MILLIS, [&]{ return emptyStateObserver.notify(); }});
+    ASSERT_TRUE(emptyStateObserver.await());
+
+    // test StateObserver on existing instance, notify() but don't await()
+
+    ASSERT_TRUE(emptyStateObserver.notify());
+}
+
+} // namespace test
+} // namespace roboticslab
